requirements.txt
fastapi
pillow
requests
uvicorn